The swampmonster took over my computer and pushed all of these changes to the server.

+++ ICE-Server/ICE-Server/Repository/BroadcastsRepository.cs
@@ -6,22 +6,26 @@ using System.Collections.Concurrent;
 using ICE_Server.Interfaces;
 using ICE_Server.Models;
 using ICE_Server.DAL;
-
+using System.Data.Entity.Infrastructure;
 
 namespace ICE_Server.Repository
 {
     public class BroadcastsRepository : IRepository<Broadcast>
     {
-        ICEContext context = new ICEContext();
-        
+        private ICEContext context = new ICEContext();
+        public BroadcastsRepository(ICEContext context)
+        {
+            this.context = context;
+        }
+
         public IEnumerable<Broadcast> GetAll()
         {
             return context.Broadcasts;
         }
 
-        public Broadcast Get(int ids)
+        public Broadcast Get(int id)
         {
-            var result = (from r in context.Broadcasts where r.ID == ids select r).FirstOrDefault();
+            var result = (from r in context.Broadcasts where r.ID == id select r).FirstOrDefault();
             return result;
         }
 
@@ -34,15 +38,41 @@ namespace ICE_Server.Repository
 
         public bool Update(Broadcast item, int[] ids)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Entry(item).State = System.Data.Entity.EntityState.Modified;
-            context.SaveChanges();
-            return true;
+
+            try
+            {
+                context.SaveChanges();
+                return true;
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                throw;
+
+            }
+
         }
         public bool Delete(Broadcast item)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Broadcasts.Remove(item);
             context.SaveChanges();
+
             return true;
+        }
+
+        private bool checkEntry(int id)
+        {
+            return context.Broadcasts.Count(e => e.ID == id) > 0;
         }
